Redwing integration patch 1 for Blackmoth.
Adds support for proper angle detection when sharp shadow is on by setting a very small dash vector which is effectively
zero but which notifies my mod which direction you are dashing in.

Also makes the blackmoth instance public so I can check when grubberfly is active and stop using my abilities.

This could be done without this but that would require awful reflection which is terribly slow.